refactor: move daemon-manager to shared, fix CORS/WS origin, apply reconnect poller
- Move daemon-manager.ts from pidash/ to shared/ (fix cross-extension coupling)
- Add callerUrl param to ensureUiBuilt for correct path resolution
- Fix CORS origin reflection — validate against localhost regex
- Fix WS origin prefix bypass — regex instead of startsWith
- Apply setupReconnectPoller to pidash.ts (was only in pidiff)
- Update AGENTS.md structure

Addresses Qodo review findings on PR #444.
