+++ AGENTS.md
@@ -73,7 +73,6 @@ pi-config/
 │   ├── pidash/                      # Live web dashboard extension (standalone)
 │   │   ├── index.ts                 # Entry point
 │   │   ├── pidash.ts                # Dashboard logic (connects to pidash daemon, forwards provider response info)
-│   │   ├── daemon-manager.ts        # Daemon infrastructure (spawn, health check, WebSocket) — shared by pidash and pidiff
 │   │   └── pidash-ui/               # React + shadcn/ui web dashboard
 │   │       ├── src/                 # React source (components, hooks, types)
 │   │       └── dist/               # Built output (generated, gitignored)
@@ -82,7 +81,9 @@ pi-config/
 │   │   ├── pidiff.ts                # Diff viewer logic (spawns/connects to pidiff daemon)
 │   │   └── pidiff-ui/               # React diff viewer UI (@pierre/diffs + @pierre/trees)
 │   ├── shared/                      # Shared extension utilities
-│   │   └── ws-client.ts             # WebSocket heartbeat + reconnect helpers (used by pidash, pidiff)
+│   │   ├── daemon-manager.ts        # Daemon infrastructure (spawn, health check, WebSocket) — shared by pidash and pidiff
+│   │   ├── ws-client.ts             # WebSocket heartbeat + reconnect helpers (used by pidash, pidiff)
+│   │   └── ui/                      # Shared shadcn/ui components (used by pidash-ui and pidiff-ui via @ui alias)
 │   ├── acpx-provider/              # ACPX provider extension (acpx/runtime library API)
 │   │   └── index.ts                # Provider + exported discoverAcpxModels() for external consumers
 │   └── image-gen/                   # Image generation extension (standalone)
